requirements.txt
psycopg2==2.7.4
markupsafe==1.0
sqlparse==0.2.4
itsdangerous==2.0.1
Flask==0.12.2
Flask-Admin==1.5.1
Flask-JSON==0.3.2
flask_login==0.4.1
flask-restplus==0.10.1
jsonschema==2.6.0
marshmallow==2.15.1
Flask-Cache==0.13.1
python-pam==1.8.3
wtforms==2.1
cherrypy==14.2.0
natsort==5.3.0
watchdog==0.8.3
dateutils==0.6.6
Werkzeug==0.16.1
Jinja2==3.0.3
